Moved build_schedule and tests
